chore: moved api/bot logic to their own controllers
