+++ src/controllers/apiController.ts
@@ -1,0 +1,22 @@
+import express from 'express';
+import path from 'path';
+import cors from 'cors';
+import config from '@root/config';
+import { router as apiRouter } from '@routers/apiRouter';
+
+export const app = express();
+app.use(express.json());
+app.use(cors());
+app.use('/api', apiRouter);
+
+const webAppPath = path.join(__dirname, '..', 'web', 'dist');
+app.use(express.static(webAppPath));
+app.get('*', (req, res) => {
+	res.sendFile(path.join(webAppPath, 'index.html'));
+});
+
+export function startApiServer() {
+	app.listen(config.PORT, () => {
+		console.log(`[API] Running on PORT = ${config.PORT}`);
+	});
+}
