+++ src/events/discordEvents/clientReady.ts
@@ -1,5 +1,5 @@
 import { type Client } from 'discord.js';
 
 export default function ClientReady(client: Client<true>) {
-	console.log(`Ready! Logged in as ${client.user.tag}`);
+	console.log(`[BOT] Connected as ${client.user.tag}`);
 }
